ms_gateway/requirements.txt
fastAPI
uvicorn
python-dotenv
httpx
requests